flask-sqlalchemy/requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
Pillow==9.0.1
pycodestyle==2.8.0
PyMySQL==1.0.2
python-http-client==3.3.7
pytz==2022.1
randominfo==2.0.2
requests==2.27.1
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.0.3
stripe==2.68.0
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.0.3
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==37.0.2
Flask==2.0.3
Flask-Cors==3.0.10
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
python-http-client==3.3.7
pytz==2022.1
randominfo==2.0.2
requests==2.27.1
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.0.3
stripe==2.68.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.0.3
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
python-dotenv==0.21.0
Flask-Migrate==3.1.0